requirements.txt
streamlit
phi-assistant
duckduckgo-search
anthropic
